Bank/requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-braces==1.13.0
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-crud-generator==0.3.6
django-crudbuilder==0.2.7
django-debug-toolbar==1.11
django-extensions==2.2.1
django-extensions-shell==1.7.4.1
django-filter==1.1.0
django-guardian==2.0.0
django-joinfield==0.2
django-rest-auth==0.9.5
django-rest-delegated-permissions==1.4.0
django-rest-swagger==2.2.0
django-tables2==2.0.6
django-taggit==0.17.1
djangorestframework==3.9.0
djangorestframework-filters==0.11.1
et-xmlfile==1.0.1
httpie==1.0.2
idna==2.8
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
jwt==0.6.1
lazy-object-proxy==1.3.1
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysql-connector-python==8.0.13
mysqlclient==1.3.14
oauthlib==3.0.2
OdooRPC==0.7.0
openapi-codec==1.3.2
openpyxl==2.5.12
Pillow==5.3.0
protobuf==3.6.1
psycopg2==2.7.6.1
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
rest-condition==1.0.3
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.0.0
Werkzeug==0.15.5
wkhtmltopdf==0.2
wrapt==1.10.11
